Nuevas Clases y Estructuras DTO

+++ src/main/java/umanizales/motogp/model/Classification.java
@@ -10,7 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Classification {
-    private Long code;
+    private int code;
     private String description;
     private List<ClassificationTime> grill;
     private boolean state;
